pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "autorag-architect"
version = "0.1.0"
description = "Autonomous RAG System Architect & Pipeline Engine"
authors = [
    {name = "AI Engineer", email = "engineer@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "langchain>=0.1.13",
    "langchain-openai>=0.1.1",
    "langchain-community>=0.0.29",
    "langchain-text-splitters>=0.0.1",
    "sentence-transformers>=2.5.1",
    "chromadb>=0.4.24",
    "pypdf>=4.1.0",
    "python-docx>=1.1.0",
    "markdown>=3.6",
    "beautifulsoup4>=4.12.3",
    "tiktoken>=0.6.0",
    "litellm>=1.34.0",
    "pyyaml>=6.0.1",
    "python-multipart>=0.0.9",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.5",
    "httpx>=0.27.0",
    "anyio>=4.0.0",
    "ruff>=0.3.4",
    "mypy>=1.9.0",
    "pre-commit>=3.7.0",
    "pytest-cov>=5.0.0",
]
cli = [
    "typer>=0.12.3",
    "rich>=13.0.0",
]
ml = [
    "mlflow>=2.12.0",
    "ragas>=0.1.0",
]

[project.scripts]
autorag = "src.cli.main:app"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "RUF"]
ignore = ["E501"]  # line length handled by formatter

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
fail_under = 60
show_missing = true
